requirements.txt
PyYAML==6.0.2
torch==2.4.1
tqdm==4.67.1
wandb==0.19.6
pillow==10.4.0
pandas==2.0.3
numba==0.58.1